releases : separate out github specific parts
